Update: #231 Wave 2 후속 — feature/home/api KMP 전환
