--- feature/home/api/src/main/java/me/pecos/memozy/feature/home/api/HomeNavigation.kt
@@ -1,22 +1,0 @@
-package me.pecos.memozy.feature.home.api
-
-import androidx.navigation.NavGraphBuilder
-
-/**
- * 홈 화면으로 진입하기 위한 네비게이션 정의
- */
-object HomeRoute {
-    const val LOGIN = "login"
-    const val MAIN = "main"
-    const val SETTINGS = "settings"
-    const val TRASH = "trash"
-}
-
-interface HomeNavigation {
-    fun registerGraph(
-        navGraphBuilder: NavGraphBuilder,
-        onNavigateToMemo: (String) -> Unit,
-        onNavigateToDonation: () -> Unit,
-        onBack: () -> Unit
-    )
-}
